moves menu nav and form from components into render folder
