+++ js/render/shared/movie-info.modal.js
@@ -55,5 +55,4 @@ function renderMovieInfoModal() {
     }
 }
 
-
 export { renderMovieInfoModal };
