+++ js/render/sign-in-page/sign-in.js
@@ -1,12 +1,12 @@
-import { formComponent } from "../../components/_form.js";
-import { menuNavComponent } from "../../components/_menu-nav.js";
+import { renderSearchForm } from "../shared/form.js";
+import { renderMenuNav } from "../shared/menu-nav.js";
 import { renderCopyright } from "../shared/copyright.js";
 import { renderSignInForm } from "./sign-in-form.js";
 
 function initSignInPage() {
-    menuNavComponent();
+    renderMenuNav();
     renderSignInForm();
-    formComponent();
+    renderSearchForm();
     renderCopyright();
 }
