+++ js/render/watchlist-page/watchlist-page.js
@@ -1,14 +1,14 @@
-import { formComponent } from "../../components/_form.js";
-import { menuNavComponent } from "../../components/_menu-nav.js";
+import {renderSearchForm } from "../shared/form.js";
+import { renderMenuNav } from "../shared/menu-nav.js";
 import { renderCopyright } from "../shared/copyright.js";
 import { renderMovieInfoModal } from "../shared/movie-info.modal.js";
 import { renderWatchlistMovies } from "./watchlist-movies.js";
 
 function initWatchlistPage() {
-    menuNavComponent();
+    renderMenuNav();
     renderCopyright();
     renderWatchlistMovies();
-    formComponent();
+    renderSearchForm();
     renderMovieInfoModal();
 }
